requirements.txt
Flask == 2.2.3
Flask-RESTful == 0.3.10
dnspython == 2.3.0
pymongo == 4.6.1
flask_cors
Werkzeug==2.2.2
jwt == 1.3.1
python-dotenv == 1.0.1
bcrypt
python-jose